backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
pydantic-settings==2.5.2
